Update dependecy-installation for multiple platforms; add progress output.

+++ install_dependencies.sh
@@ -1,2 +1,11 @@
-apt-get update -qq
-apt-get install -qq libgsl0-dev libarmadillo-dev libboost-all-dev
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    if [[ -f /etc/debian_version ]];then
+        apt-get update -qq
+        apt-get install -qq libgsl0-dev libarmadillo-dev libboost-all-dev
+    elif [[ -f /etc/centos-release ]]; then
+        yum -y install boost-devel armadillo-devel gsl-devel
+    fi
+
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install gsl armadillo boost
+fi
